Update cprograms.c
Changed integer to floating point variable.

+++ cprograms.c
@@ -1,10 +1,10 @@
 #include <stdio.h>
 int main()
 {
-    int cel,far;
+    float cel,far;
     printf("Enter the celsius ");
-    scanf("%d", &cel);
+    scanf("%f", &cel);
     far=cel*9/5+32;
-    printf("Fahrenheit value is %d", far);
+    printf("Fahrenheit value is %f", far);
 }
-//Refer - https://onecompiler.com/c/43ze3pwub
+//Refer - https://onecompiler.com/c/43ze3xwr9
